comments-service/requirements.txt
Django~=6.0
requests~=2.32.5
psycopg2~=2.9.11
djangorestframework~=3.16.1